Added `maps` to the `DESCRIPTION` file to facilitate the use of the `plot` method.
Set up plotting using the `maps` package and added a by_db option, although it's not actually implemented yet.

Have worked through some of the `Rmd` stuff, but most of this has been developed without internet (!) so some of it might not compile properly.

`get_occurrence` is supposed to be the generalized version of the `get_*` functions.  Cleaned up the `items` in the return component.

Cleaned up the documentation fo `get_by_taxon`.  Still running into an issue with the `record_cols`, but I'll try to clean that up as well.

Added the `NA` block to `get_by_bbox`, this should reliably add NA columns to the data responses.

Same block added to `get_by_age`.

`README` changes extend the description of the package & change text in the `Example` block.